refactor(vwegolf): split KCAN TX/actuation path into vwegolf_tx.cpp
vehicle_vwegolf.cpp had grown to 1123 lines carrying both the RX/decode
path (IncomingFrameCan2/Can3, Ticker1) and the entire KCAN TX/actuation
path in one translation unit. Two forks are actively editing this file,
and because the TX functions sit at the tail, recent merges with upstream
master have produced mechanical moved-function-body conflicts that have
to be resolved by hand even when neither side changed any logic.

Splitting the TX side out gives the two paths independent conflict
surfaces. It also puts the actuation code in its own file ahead of
charging control, which lands entirely on the TX side.

Moved verbatim: CommandLock, CommandUnlock, CommandMirrorFoldIn,
CommandHorn, CommandIndicators, CommandPanic, WakeKcanBus,
CommandWakeup, SendOcuHeartbeat, SendNmAlive, SendClimaBapBurst,
CommandClimateControl. The "Vehicle commands" section banner travels
with them, since it labels exactly that block.

There is no functional change. The moved function bodies are
byte-identical to their previous contents and the retained region of
vehicle_vwegolf.cpp is byte-identical to the previous file's first 811
lines. Beyond that relocation the diff contains only: the new file's
license header, its `#include "vehicle_vwegolf.h"`, its `#undef TAG` /
`#define TAG "v-vwegolf"` pair (TAG is a per-file macro, so the new
translation unit needs its own definition to log), one srcs line in
CMakeLists.txt, one SRC line in tests/Makefile, and one comment in
tests/test_clima.cpp that named the old file as SendClimaBapBurst's
home. component.mk needs no change — it globs src/*.cpp.
vehicle_vwegolf.h is unchanged. Native tests: 88/88 pass.
