+++ vehicle/OVMS.V3/components/vehicle_vwegolf/CMakeLists.txt
@@ -3,6 +3,7 @@ set(include_dirs)
 
 if (CONFIG_OVMS_VEHICLE_VWEGOLF)
   list(APPEND srcs "src/vehicle_vwegolf.cpp")
+  list(APPEND srcs "src/vwegolf_tx.cpp")
   list(APPEND include_dirs "src")
 endif ()
