+++ vehicle/OVMS.V3/components/vehicle_vwegolf/tests/Makefile
@@ -8,6 +8,7 @@ CXXFLAGS := -std=c++17 -Wall -Wextra -g \
 
 SRC := mock/mock_ovms.cpp \
        ../src/vehicle_vwegolf.cpp \
+       ../src/vwegolf_tx.cpp \
        test_can_decode.cpp \
        test_crtd_replay.cpp \
        test_clima.cpp
